requirements.txt
mysql-connector-python==8.2.0
pythonping==1.1.4
pyyaml>=6.0
speedtest-cli>=2.1.3
